requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
pillow==10.1.0
requests==2.31.0
sqlalchemy==2.0.23
faiss-cpu==1.7.4
numpy==1.26.2
opencv-python==4.8.1.78
pytest==7.4.3
pydantic==2.5.2
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
aiofiles==23.2.1